move proxy errors to common contract
