+++ protocol/contracts/proxy/src/lib.rs
@@ -19,6 +19,7 @@
 pub mod proxy {
 
     use common::err;
+    use common::common::Error;
     #[allow(unused_imports)]
     use ink::env::debug_println as debug;
     #[allow(unused_imports)] // do not remove StorageLayout, it is used in derives
@@ -27,18 +28,6 @@ pub mod proxy {
     #[ink(storage)]
     #[derive(Default)]
     pub struct Proxy {}
-
-    /// The errors that can be returned by the Proxy contract.
-    #[derive(PartialEq, Debug, Eq, Clone, Copy, scale::Encode, scale::Decode)]
-    #[cfg_attr(feature = "std", derive(scale_info::TypeInfo, StorageLayout))]
-    pub enum Error {
-        NotAuthorised,
-        TransferFailed,
-        SetCodeHashFailed,
-        InvalidDestination,
-        UnknownMessage,
-        NotAuthor,
-    }
 
     #[derive(PartialEq, Debug, Eq, Clone, Copy, scale::Encode, scale::Decode)]
     #[cfg_attr(feature = "std", derive(scale_info::TypeInfo))]
